DEV-1210 Remove unnecessary declaring of queue
When FxpConsumer finalizes the handling of a message (this means
success and exception) it sends a FxpResponse to a queue. Before
sending out said response, it creates a new channel in order
to declare the queue. It does so for every message and that channel
is never closed. However, declaring the queue is not needed at all,
as it should already exist. As a result, no new channel will be opened.